updating the index.js with test code and restructuring the module folders to keep commonUtils separate and FF1 core code separate
