+++ FF-1/CharMap.js
